blog: rework the global-virtual-store post; share the TOC star nudge; deterministic blog-index date sort
